Fixed some 64-bit-ificataion.

+++ Frameworks/OWF/Content.subproj/OWCompoundObjectStreamCursor.m
@@ -1,4 +1,4 @@
-// Copyright 1997-2005 Omni Development, Inc.  All rights reserved.
+// Copyright 1997-2005, 2011 Omni Development, Inc.  All rights reserved.
 //
 // This software may only be used and reproduced according to the
 // terms in the file OmniSourceLicense.html, which should be
@@ -101,8 +101,7 @@ RCS_ID("$Id$")
     [[cursors objectAtIndex:cursorIndex] ungetObject:anObject];
 }
 
-- (unsigned int)seekToOffset:(int)offset
-  fromPosition:(OWCursorSeekPosition)position;
+- (NSUInteger)seekToOffset:(NSInteger)offset fromPosition:(OWCursorSeekPosition)position;
 {
     if (abortException)
         [abortException raise];
